requirements.txt
flask
python-telegram-bot==20.8
yookassa
	